requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Pillow==10.4.0
WTForms==3.1.2
Werkzeug==2.3.8
itsdangerous==2.1.2
Jinja2==3.1.3
